Lookup peers in the client pool by ID instead of by IP address
Upstream geth uses a client pool to determine which peers to allow
and stay connected to and which not. It is supposed to reward good
peers who only stay connected as little as necessary and punish
peers who hog the connection. However, since there is nothing that
can truly uniquely identify clients (enode addresses are trivially
changable), upstream geth uses IP addresses as a proxyy. However this
breaks light clients with the same IP addresss, as well as setups like
Kubernetes that appear to geth under the same IP address due to their
networking.

Thus this change is a short term hack to stabilizes connections for us
by doing the lookup by the enode ID. Longer-term we need to figure out
a solution to this. Check out https://github.com/celo-org/celo-monorepo/issues/522
and https://github.com/ethereum/go-ethereum/pull/16899#issuecomment-418513423