Handle crash during rollback in CeloLatestSync mode
We don't have the full header chain in the CeloLatestSync mode,
therefore, a rollback beyond oldest fetched block will lead to a crash.
An example is described in https://github.com/celo-org/celo-monorepo/issues/1728#issuecomment-447154687

To prevent that, in CeloLatestSync mode, we do not roll back beyond the
oldest block. Note that since we fetch last 128 blocks, this will
interfere with a large scale re-org of the chain. In a PoA network, we
should not be expecting a re-org greater than 128 blocks anyways.