Creating a README.celo.md
README.md is already taken, so, I am using README.celo.md to add Celo-specific information. Some of this was in the Run book - https://docs.google.com/document/d/1vMXZY6pu2fmwHxTzZE0YbKhn6Gs9szPZjHtjvV8dsis/edit# but it is not always obvious to look into the runbook for that, therefore, I am adding the information in this file.