Trianglesphere/fixed tests (#130)
* Merges upstream doTest changes to ci.go

This adds changes that were made up upstream to remove `go vet` from
running during the test phase as the lint phase includes `go vet` as
well as more lint checks.

* Make circleci and local testing consistent

* Switches from make to ci.go

`make test` depends upon the `all` target, which does an install. I
think that the install is causing issues with memory on circle ci.

* Adds guard on test to stop it from running when it will fail

The TestCLISwarmFs fails in the circle CI environment, and this should
fix it.