Make Istanbul's Seal asynchronous
Originally, when Istanbul was written, the Seal method was synchronous,
i.e. it would return when the sealing process was done (or cancelled).
Now that is no longer the case, in fact the miner expects the Seal to
return to be able to respond/intiate other Seal's. This commit allows
Istanbul to do so.