[Istanbul] Seal/NewHeadChain fixes (#147)
* Make Istanbul's Seal asynchronous

Originally, when Istanbul was written, the Seal method was synchronous,
i.e. it would return when the sealing process was done (or cancelled).
Now that is no longer the case, in fact the miner expects the Seal to
return to be able to respond/intiate other Seal's. This commit allows
Istanbul to do so.

* Notify the Istanbul engine of new ChainHeads

During the merge, we removed this crucial line, where the worker will
notify the Istanbul engine of new chain heads that crucially allow its
peers to increse the sequence number for the next Seal.