Merge istanbul into (celo) master (#135)
* cmd, consensus, eth, ethstats: add protocol interface into consensus to support custom messages

* params: add Istanbul consensus engine config to ChainConfig

* cmd/*: add Istanbul command line flags

* consensus/istanbul, eth: add Istanbul configuration

* node: add an interface to retrieve node key from config

* core/types: add Istanbul specific block hash calculation

* internal/web3ext: add Istanbul JS RPC API

* consensus: add Istanbul consensus engine interface

* consensus/istanbul: Istanbul core implementation

* consensus/istanbul: add tests for Istanbul core

* consensus/istanbul: common Istanbul interfaces and types

* consensus/istanbul: Istanbul validator set implementation

* consensus/istanbul: Istanbul consensus backend implementation

* core, les, eth, miner: Istanbul consensus integration

* cmd/*, core, params: add ottoman testnet

* Address comments

* Fix Istanbul lint (#137)

* Remove unused vars for lint

* Remove unnecessary conversion

* Don't propose empty blocks with Istanbul (#134)

* Make istanbul.Backend a public type

* Don't mine empty block with Istanbul engine unless necessary

* PR comments

* Revert to original order

* Add test fixes

* PR comments

* Update the expected hash given that celo headers have a different payload with BlockSignature

* Make Istanbul's Seal asynchronous

Originally, when Istanbul was written, the Seal method was synchronous,
i.e. it would return when the sealing process was done (or cancelled).
Now that is no longer the case, in fact the miner expects the Seal to
return to be able to respond/intiate other Seal's. This commit allows
Istanbul to do so.

* Notify the Istanbul engine of new ChainHeads

During the merge, we removed this crucial line, where the worker will
notify the Istanbul engine of new chain heads that crucially allow its
peers to increse the sequence number for the next Seal.

* Apply celolatest syncmode mods to Istanbul

* [Istanbul] Seal/NewHeadChain fixes (#147)

* Make Istanbul's Seal asynchronous

Originally, when Istanbul was written, the Seal method was synchronous,
i.e. it would return when the sealing process was done (or cancelled).
Now that is no longer the case, in fact the miner expects the Seal to
return to be able to respond/intiate other Seal's. This commit allows
Istanbul to do so.

* Notify the Istanbul engine of new ChainHeads

During the merge, we removed this crucial line, where the worker will
notify the Istanbul engine of new chain heads that crucially allow its
peers to increse the sequence number for the next Seal.

* Fix test

* Fix test