Removes src validator from handler functions
Every message is signed by the validator that sent the message. That
signature is checked against the address in the message, and thus the
address of the message is the original sender of the message. In the
case that extra data around a validator is needed (when checking BLS
signatures), the validator can easily be looked up in the validator set.

This also changes a lot of tests to work with these changes.