Fixes some test flakiness and error reporting
For CommitsBlocksAfterRoundChange, because no single node committed a block
before other nodes went into a round change, there is no guarantee on which
block will be committed other than that block needs to be consistent.

For PreparedCertificatePersistsThroughRoundChange I'm finding that it will
occasionally have a node that does not commit the latest block that every
other node has committed when the test times out or ends.