Move Proposer outside ValidatorSet
Now ValidatorSet is inmutable, meaning that remains the same for the
whole epoch. While, before ValidatorSet changed every roundChange since
it had the proposer.

Also, by doing this, we eliminate a way of changing data inside the
RoundState without calling any RoundState function.

RounState has a MessageSet which has a ValidatorSet

And on some scenarios the ValidatorSet proposer was being changed
withouth notifying the RoundState