Move Proposer outside ValidatorSet (#628)
Now ValidatorSet is immutable, meaning that remains the same for the
whole epoch. While, before ValidatorSet changed every roundChange since
it had the proposer.

Also, by doing this, we eliminate a way of changing data inside the
RoundState without calling any RoundState function.

RoundState has a MessageSet which has a ValidatorSet

And on some scenarios the ValidatorSet proposer was being changed
without notifying the RoundState