Persist RoundState on every transition (#632)
Active Validators have to remember their state in the ibft consensus
protocol. For that we need to persist state after each transition.

To enable that, we create a new RoundState implementation that serves as
a proxy to the existing one. This proxy will store the state after each
mutation, into a levelDB database.

To save RoundState into the levelDB database, we need to serialize it.
We use RLP encoding for that, and several new Encode/Decode functions
were needed:

 - Validator
 - ValidatorSet
 - MessageSet
 - RoundState

 Custom encode/decode logic is required to support structs with
 interface fields and nils.