Add garbage collection for RoundStateDB
To implement this we had to change the key format on levelDB; so that
Range queries over istanbul.View are possible.

To do so, a View is serialized to a byte[] as 2 uint64 (pair sequence,
round)

Also, garbage colleciton is optional and is turned on using constructor
options within the newRoundStateDB()

Finally, to implement a recurrent task, a new common/task package was
created. This makes it easy to test the recurrent task logic and it can
be also used in many other places since it's a quite commom patter
within geth.