IBFT: Add ParentSeal on Finalize (not in Prepare)
Refactor istanbul engine so that the "ParentSeal" is added to the header
in the Finalize Step, instead of the Prepare step.

It extracts logic to it's own function to make it more readable, and
switched to a style of `if err => return` instead of relying on nested
ifs.

Minor behaviour change: There are 4 main outputs for the addParentSeal
now

 1. block number <=1 => do nothing
 2. failed to fetch parentHeader.AggregatedSeal => return error
 3. failed to generate union of seals => use parentHeader.AggregatedSeal
 as ParentSeal
 4. All ok => use UnionOfSeal as ParentSeal

 Before this commit, when it failed to fetch a seq it would do nothing
 but should be case (3)