Remove eth63/64 in favor of istanbul64/65
We don't use eth variant. The only variant we did use use istanbul64.

With 1.9 there's a new protocol eth/64 that collides with istanbul/64.
So that protocol was renamed to istanbul/65.

istanbul/64 is eth/63 + istanbul messages.

This commit eliminates the other protocols, and makes istanbul variants
the only protocols we support.

We also "Fix" the EthProtocolVersion() rpc. Since we know return our
protocol version; not ethereum's.

There was also some support for eth/62 that we've never used, so that
was also removed.