Fix Bad detection of empty ouptu on contract calls
1.9 fails differently on `return empty`. Which is what Proxy.sol does
when no implementation has been set.

Instead if fails with "attemping to unmarshall empty arguments".

This commits indivualized the error and uses it to detect this specific
registry error, as master branch does.

A proper fix is to change Proxy.sol implementation which has been done
already, but it's not backward compatible with baklava.