Prevent ethstats from blocking ChainHead publishing (#840)
Problem.

I have a testnet with 3 validators, validator0 is proxied. Due to 1 second block times, when validator0 is just behind, it wouldn't be able to participate in consensus unless it imports blocks that brings it back to the most recent sequence. Eventually however, it would stall and be stuck at a block. Sometimes it would recover only to stop again.

Investigation:
I first tried to figure out why I would no longer import blocks. Turns out that I eventually hit this line that prevents DOS from our peers.

It turns out that something prevented the completion of an insert which let the queue accumulate. After a lot of log lines, it turns out that sending ChainHeadEvents was blocking. This can happen if one of the subscriptions channels is not accepting any messages.

After digging deeper, I realized that the ethstats subscription is the culprit. The channel has a capacity of 10, but it wasn't getting pulled from.

IsProxiedValidator actually checks whether we currently have a proxy peer. If that is not the case, we end up stuck in this else branch which itself contains a for loop out of which we never get out until there is a reporting failure.

Because we do not land in the other branch, we will never consume from signCh. Unluckily, that is exactly where the HandleLoop is trying to publish to.

By being blocked there, we no longer have a goroutine consuming from the ChainHeadEvent channel, which causes the sending of the ChainHeadEvent to block, which causes all subsequent imports of blocks to eventually stop (after 64 of them), which causes the istanbul engine to no longer be able to get to the most recent sequence.

Solution:
The ethstats code in general isn't quite pretty and since I'm focused on the 1.9 merge, I just made the publishing to the signCh non blocking, since stats paylods should be easily ignored. I also moved the defer to close the subscription into the goroutine that is actually listening in case we are breaking out of the HandleLoop for any reason.