Higher trie limit (#845)
* Increase the tries in memory limit to 1024

* Fix TestLowDiffLongChain

This was odd. Since `TriesInMemory` was used to construct a long chain,
with this PR, we massively increased the limit which caused the
diffculty of the canonical chain to be so large that the fork would not
apply. When I shortened the chain to the original length, the test
passed, despite the total difficulty of the fork still being shorter
than the canonical. I'll leave this for now and expect to revert when
we reduce the limit again

Co-authored-by: Trevor Porter <trkporter@ucdavis.edu>