p2p/discover: add initial discovery v5 implementation (#20750)
celo-blockchain cherry-pick conflicts:
  cmd/devp2p/discv4cmd.go
  p2p/discover/common.go
  p2p/discover/v4_udp.go

This adds an implementation of the current discovery v5 spec.

There is full integration with cmd/devp2p and enode.Iterator in this
version. In theory we could enable the new protocol as a replacement of
discovery v4 at any time. In practice, there will likely be a few more
changes to the spec and implementation before this can happen.