Don't commit previous sequence while on seq 0
When on the first round of consensus, sending an old commit message can
cause an underflow when looking for the validators of the parent block
of the current block. The parent block does not exist for the genesis
block, therefore when this code path is triggered before the first
commit, malicious nodes could send a commit for the n = 0 block.