Fix use of potentially invalid headers to verify new headers
VerifyHeaders() is used to verify new headers in batch obtained from peers.
To do the verification it uses 1..n headers in batch to verify header
n+1 in the batch.

The problme is that if some of if 1..n header are invalid, it will still use them
to the verification. This can be potentially be used to attack a node's
validatorSet cache for the next epoch.

This commit fix this, by failing fast. When the first invalid header is
encountered, all subsequent headers on the batch will fail with `ErrUnknownAncestor`