Update Alfajores genesis configuration to match reset network (#1118)
### Description

On July 16th, the Alfajores network was reset to bring the chain up to date with the 1.0.0 client.
This is expected to be the last such update and as a result, we want to encode the genesis block and bootnodes into the source such that a new user can simple run `geth --alfajores` and sync to the network! 🙌

Included in this PR is the configuration for the reset Alfajores network, and the removal of code required to make the client compatible with the old Alfajores network.

### Other changes

* Update `README.md` to utilize the fact that fast sync now works.

### Tested

Created a fresh node and by running `geth --alfajores` and confirmed that it could connect to the Alfajores network and begin syncing.

### Backwards compatibility

Client will no longer support networks created by pre-1.0.0 networks. This is fine because those networks no longer exist.