Adds metrics for missed rounds as proposer & downtime events (#1091)
Two sets of metrics are added.

Adds meter for the number blocks that could have been proposed by the validator but were not. This is not the same as being down for a block.
Adds a gauge for the number of blocks that were not signed by the validator in a row when the validator is elected.
Adds a meter for blocks that will be counted as downtime events & also reports an error the the logs.
The gauge for missed blocks is reset at each epoch to match how downtime is calculated.

The meter in 1 is complex because in a single sequence (block number), if there are multiple rounds, the proposer is rotated. In the first round, the proposer can propose any block, but in later rounds, if the proposer receives a prepared certificate during round change messages, it must re-propose that block. Thus from the finalized block, it is hard to tell if the proposer did not propose, proposed their own block that timed out, or were required to proposed a previous block. The meter 'blocksMissedRoundsAsProposerMeter' includes all three cases