Fix IstanbulOverride decoding error when trying to use config.toml (#1142)
### Description

Yesterday I used the `dumpconfig` command to generate a `config.toml` file and tried to use it to run a Geth node with `geth --config config.tml`, but got a decoder error. This PR includes a fix for the error in that flow.

### Other changes

* Sorted some methods in `web3ext.go` alphabetically

### Tested

Ran a node locally with the generated `config.toml`