Fix failing tx fetcher tests.
The indexes used in these tests are fixed according to the deterministic random shuffles done (only for the tests) by TxFetcher.  Since Celo transactions have additional fields, their hashes are different, leading to different orderings before and after the shuffle, so the indexes had to be updated to match the tests' intents.