eth_call and eth_estimateGas: use gas price minimum of zero and allow gas price of zero.
Previously, we were using a minimum of zero for eth_estimateGas, but not for eth_call.  As a result, calling read-only functions would require a balance in the currency used for fees.  There was an upstream hack which set the balance to maxuint, but it has been removed upstream and merged into celo-blockchain (see the previous commit).  And that hack did not work for non-CELO currencies such as cUSD.

This commit also changes how the gas price is set to 0 for gas estimation.  Instead of an extra argument to DoCall(), we now set args.GasPrice to nil in DoEstimate().