Reapply chain repair fix changes to `eth/sync.go`.
For #1163, these changes had to be adapted because we had an older version of the `eth/sync.go`.  The commit before this one merged in the upstream changes that were missing, and which overwrote the adapted chain repair fix changes.  So this commit reapplies those changes from the upstream chain repair fix commit, since they now apply as is without any issues.