Fix TestCustomGenesis() passing when it shouldn't.
Similar to the upstream fix in the previous commit, we had a problem where if the init step in this test fails, geth then defaults to using the mainnet genesis, and the test would pass because the mainnet genesis timestamp is 1587571200, which passes a regex check for "0".  This commit fixes that by changing the timestamps in the custom genesis data and the checks to longer numbers, which are not substrings of "1587571200".