Make LightPeers default to 100 again.
The change to 99 was due to an error with light peers being >= max peers.  This is no longer an issue because (a) max peers default is now 175, (b) the logic has been fixed which increases max peers accordingly if the light server is running and using the default LightPeers.  Also note that the CLI docs still said the default is 100, so with this change the CLI docs become correct as well.