Change light.serve default back to 0.
It was 0 in go-ethereum but was changed in Celo long before mainnet, with the idea that it would be better for light clients if nodes could serve them by default. However, it has since become apparent that serving light clients is just one reason to run a node (others are validators, proxies, full nodes to service the attestation service or internal services), and it makes more sense to have the light server be opt in than opt out.  Also, in our light server docs we specify --light.serve 100, so even there we are not relying on the default being 50.