Bump up the CI resource_class for unit tests and coverage. (#1269)
This changes them to use medium+ instead of the default medium, which should help with flaky tests where the machine is overloaded and which don't happen on dev machines since they're more powerful (e.g. #629). If the failures keep happening, we can either (a) increase further to large, (b) increase the waits in the tests, (c) make those specific tests not run using `t.Parallel()`.