Refactor uptime computation to use windows
We define a `Window` type and then uptime block are the sum of all
blocks where block is within monitoredWindow, and lastSignedBlock for
the validator is within currentLookback window.

Additionally, updateUptime() function now receives the block is to
monitor and the corresponding singatures, instead of receiving a block
and the parent block signatures. So math is simplified.

MonitoringWindow() now returns the blocks whose signatures we want to
monitor and not the blocks whose parentBlock signatures we want to
monitor.