Cleanup validator uptime computation (#1270)
* Rename "scoreTally" to "UptimeMonitorWindow"

Instead of one method for first and last block on window, have a method to get the window [first, last] pair.

 - Add godocs to a few utils.go methods
 - Add tests to a few utils.go methods

* Extract uptime to its own package

- Centralize all uptime related logic into istanbul/uptime pkg.
- Use interface for storage to ease testing later

* First pass on PR comments

* Refactor uptime computation to use windows

We define a `Window` type and then uptime block are the sum of all
blocks where block is within monitoredWindow, and lastSignedBlock for
the validator is within currentLookback window.

Additionally, updateUptime() function now receives the block is to
monitor and the corresponding singatures, instead of receiving a block
and the parent block signatures. So math is simplified.

MonitoringWindow() now returns the blocks whose signatures we want to
monitor and not the blocks whose parentBlock signatures we want to
monitor.