Fix strict greater than check in `canPayFee()`.
In order to be able to pay the fees for a transaction, you have to have at least the amount of the fee, not strictly more than that amount.

This was leading to spurious errors during gas estimation.  For gas estimation, the gas price is set to zero, and so `canPayFee()` should always return true, even if the account has no funds.  However, it was returning false if using CUSD as the fee currency.  If using CELO as the fee currency, it was returning true, but that is due to the upstream hack which sets the account's balance to maxuint (and which has been changed in upstream v1.9.13 and we will merge in PR #1251).