CIP-28: Split etherbase into validator & txFeeRecipient addresses (#1195)
Implementation for CIP-28.

We separate the coinbase address role into two roles: validator and txFeeRecipient. The validator address will receive validation rewards, while the txFeeRecipient address will receive transaction fees. The terms etherbase and coinbase become aliases for the txFeeRecipient address.

Both parameters are set via command line
New flags:
miner.validator is the validator address
txFeeRecipient the address where all the fees payments will be transferred
If one of these flags is present, the other will be required, even if it is the same address, to force a conscious decision about it

To maintain backwards compatibility, the etherbase flag if is already set, and non of the other two flags were set, it will fill that address as the miner.validator and the txFeeRecipient. The flag will eventually be deprecated.

We changed the default behaviour of the etherbase flag. Now it won't assume the first account of the keyStore as default. So beware. Either the etherbase flag or the pair miner.validator and txFeeRecipient MUST be set.

Backwards compatibility
Hardfork Change. Modifies block validation logic. Part of Donut Hardfork

Co-authored-by: Lucas Ege <lucas@celo.org>
Co-authored-by: Andres <andres@Andress-MacBook-Pro.local>
Co-authored-by: Gaston Ponti <pontigaston@gmail.com>