Fix indeterministic test TestSideLogRebirth (#1339)
This test from go-ethereum includes code to make the main and side-chain have difficulties as desired for the test to work.  Since for Celo the difficulty of each block is 1, that part wasn't working and the main and side-chain both had difficulty 3.  As a result, whether or not a reorg was done when the side-chain is inserted was indeterministic (see writeBlockWithState() in core/blockchain.go).  But Go's math/rand global rand object uses the seed value 1, and so you get the same sequence of values every time.  This was leading to some very confusing behavior: this test was passing as part of the overall test suite, but failing if run by itself.  This commit fixes that by making the main chain have 3 blocks, the side chain 2 blocks, and then it adds 2 more to side chain, making it 4.  This way, the result is always as desired: there is no reorg when the side-chain is constructed and there is a reorg at the last step when we add 2 more.a reorg at the last step when we add 2 more.