Lift up methods from evm (incomplete)
With this commit, we start the refactor. We have evm layer, and
contracts layer. The ojective is to decouple evm from contracts,
and let it just be a clean evm more similar to go-ethereum one

1) We introduce a new flavour of contract function. There's the
contract_com/evm.go and thew new contracts/evm.go

The first, is now a backward compatiblity layer that just calls the new
function on evm.go.

Before, every function would call `createEVM()` and then use the EVM to
make a call. Instead, now on contract_comm we call the
`defaultCallerFactory` that returns a `ContractCaller`. This contract
caller is the parameter we then use to call a function with the same
name on the `contracts/evm.go` module.

A `ContractCaller` is just an interface with single implementation that
behaves like a thin wrapper around an EVM instance. So, in essence we
are just using the same logic, with more indirection.

The caller implementation lives on the vm package (for now) and the factory lives on
the contracts/caller/evm package.

The rationale for this change, is that we want to eventually remove th
need of a singleton, and instead move to a instance based contract
calling. Additionally, we need to introduce an interface to break a
circular dependency with the EVM (evm call contracts, contracts call evm).

2) We move methods from the EVM to the `contracts` package
 - StaticCallFromSystem
 - CallFromSystem
 - handelABICall

These methods were introduced by us to make calling contracts simpler,
but they are better places on a higher layer, not in the EVM.

3) We move the GetRegistryWithEVM from `vm` package to `contracts`

Idea here is the vm module should not be aware of contracts, nor do
contract calls.

FINAL REMARKS:

This is not the final state, but a first step in a refactoring journey.
The goal is to decouple EVM from contracts, and clean it. While at the
same time improving how we call contracts.

WIP

Fix: Missing init of internalEVMHandler

Fix: Calls SetDefaultCallerFactory where needed