Stop verifying individual commit messages
Since we now want to verify commit messages in aggregate. We cannot rely
on commit messages until we have have a quorum of commit messages with a
valid aggregate signature.

This change removes the individual verification of commit messages and
removes reliance on the commit messages, except for when we verify
the signature in aggregate. Previously commit messages were being
counted for the transition from preprepared to prepared, this is no
longer the case. As such the code to handle the preprepared -> prepared
transition during the commit phase is no longer needed.

In the case of commit messages for the previous sequence they were
previously being aggregated when the engine prepares the header for the
next sequence and if the aggregate signature failed verification during
this we would fall back to just using the aggregate signature of the
previous block. This has not changed although now that the commit
messages are not being individually verified it opens the possiblilty
that the aggregate signature is invalid, which looks like it shouldn't
have happened before, even though the code allowed for it.