Removes evm dependency from contracts
EVM requires two things obtained from the contract layer to operate

 * compute tobin tax on any transfer
 * verify caller is goldtoken address in transfer precompile

 Before, we relied on contracts module (before that contract_comm).

 Now, we rely on evm.Context to expose the necessary methods.

Additionally, from EVM perspective there's no tobin tax, just a Transfer
function on the evm.Context. Tobin tax is just the part of the transfer
function implementation that is defined when creating a evm.Context

With this change vm.NewContext requires an extra paramenter
ContractsContext that provided the 2 required functions. To ease the
task of calling this, now there's a contract.NewContext() with old
signature that is now used for every previous user of vm.NewContext()