Check aggregate seal for current seq on commit
Moved aggregate seal generation out of backend and into core, aggregate
seals are now verified in core.

Since we are now using the aggregate seal as a means to verify the
validity of received commit messages it makes sense to do the
verification in core, where previously we were verifying individual
signatures.