Fix panic with discv5
This is due to a change in golang 1.16 which panics when using a non-unique
hash in a HMAC. This change copies the upstream change which uses a new
sha instance rather than the existing hash value.

Discv5 is next touched in: https://github.com/ethereum/go-ethereum/pull/21647
where it is reworked.

Full history of the encoding test:
git log geth/master --full-history -- p2p/discover/v5_encoding_test.go