Manage tx nonces locally for load-bot
By default the ethereum tooling relies on PendingNonceAt function to get
the nonce for sent transactions. PendingNonceAt looks at the tx pool to
derive the nonce but there is no guarantee that the pool is up to date
with sent txs, especially when under heavy load. Relying on
PendingNonceAt results in attemts to use the same nonce twice which
breaks the load-bot.

This change instead uses PendingNonceAt on startup only, and manages
nonces locally from then on.