EVM: Removes static call to GetRegisteredAddress
This commit removes the use of a static call to obtain a registry
address. This was used on the Transfer precompile to check that the only
caller to the contract is the GoldToken contract.

With this commit, there's a GetRegisteredAddress function on the
evm.Context that is injected at the moment of creating.

For now, this function is hardcoded, but later it will be received as a
parameter to the vm.NewContext()