Refactors contracts calls
1) Defines Contract with a Query (static call) and Execute (call)
methods.

2) Replaces evm by the Backend interface which will allow a
caching decorator implementation later

3) Simplifies code by adding a little duplication but a more linear code
flow

4) Replaces header & state by evm parameter

In order to isolate us from header & state, we initially use the evm

5) Replaces evm for contracts.Backend in all fun signatures

We need to remove the dependency to evm and replace it for an interface.
This will allow us to provide a different implementation that add a call
caching feature.

This commit, replaces the evm parameter by a ContractBackend parameter
in all contract_comm function signatures.

In order to do so, we added a few methods to the interface so that
contracts.GetRegistryAddress can work. And replaces direct acces to
DontMeterGas field by equivalent function calls.

6) Adds Backend that ensures a single evm per call