Replaces evm for contracts.Backend in all fun signatures
We need to remove the dependency to evm and replace it for an interface.
This will allow us to provide a different implementation that add a call
caching feature.

This commit, replaces the evm parameter by a ContractBackend parameter
in all contract_comm function signatures.

In order to do so, we added a few methods to the interface so that
contracts.GetRegistryAddress can work. And replaces direct acces to
DontMeterGas field by equivalent function calls.