Uses block context within txpool
We use txpool.reset() which is the moment on which the txpool updates to
a new blockchain HEAD; to create a blockContext. The reset was already
being used for:

 * removes txs mined on previous blocks
 * updates block max gas limit
 * upddates pool.currentState

So it's only natural to create a block context there.

Thanks to doing this we avoid calls for every tx on:
 * gaspriceminimun for every tx
 * gas for alternative currency
 * whitelisted currencies